# backend/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Auction System",
        "description": "Auction System API Documentation",
        "version": "2.0.0"
    },
    "paths": {
...
        "/api/v1/notifications/": {
            "get": {
                "tags": [
                    "Notifications",
                    "Notifications"
                ],
                "summary": "Get Notifications",
                "description": "Get user's notifications with pagination\n\nGET /notifications?skip=0&limit=50\nHeaders: Authorization: Bearer \u003Caccess_token\u003E\nReturns: Paginated list of notifications",
                "operationId": "get_notifications_api_v1_notifications__get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "skip",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 0,
                            "default": 0,
                            "title": "Skip"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1,
                            "default": 50,
                            "title": "Limit"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": true,
                                    "title": "Response Get Notifications Api V1 Notifications  Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/notifications/auction/{auction_id}": {
            "get": {
                "tags": [
                    "Notifications",
                    "Notifications"
                ],
                "summary": "Get Auction Notifications",
                "description": "Get notifications related to specific auction\n\nGET /notifications/auction/{auction_id}\nHeaders: Authorization: Bearer \u003Caccess_token\u003E\nReturns: List of auction-related notifications",
                "operationId": "get_auction_notifications_api_v1_notifications_auction__auction_id__get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "auction_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Auction Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": true,
                                    "title": "Response Get Auction Notifications Api V1 Notifications Auction  Auction Id  Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/sse/auction/{auction_id}": {
            "get": {
                "tags": [
                    "SSE",
                    "Real-time Updates"
                ],
                "summary": "Stream Auction Updates",
                "description": "Stream real-time auction updates via SSE\n\nGET /sse/auction/{auction_id}\nHeaders: Authorization: Bearer \u003Caccess_token\u003E\nReturns: Server-Sent Events stream with formatted auction data\n\nEvent types:\n- auction_update: Regular auction state updates (every 2s)\n- heartbeat: Keepalive signal (every 30s if no data change)\n- error: Error messages\n\nFixed: Database session leaks, NoneType crashes, proper error handling",
                "operationId": "stream_auction_updates_api_v1_sse_auction__auction_id__get",
                "parameters": [
                    {
                        "name": "auction_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Auction Id"
                        }
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Token"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/sse/auctions/active": {
            "get": {
                "tags": [
                    "SSE",
                    "Real-time Updates"
                ],
                "summary": "Stream Active Auctions",
                "description": "Stream real-time updates for all active auctions via SSE\n\nGET /sse/auctions/active\nHeaders: Authorization: Bearer \u003Caccess_token\u003E\nReturns: Server-Sent Events stream with formatted data for all active auctions\n\nUpdates every 5 seconds with current state of all active auctions\nFixed: Database session leaks, N+1 queries, batch processing",
                "operationId": "stream_active_auctions_api_v1_sse_auctions_active_get",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Token"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/sse/notifications": {
            "get": {
                "tags": [
                    "SSE",
                    "Real-time Updates"
                ],
                "summary": "Stream User Notifications",
                "description": "Stream real-time notifications for current user via SSE\n\nGET /sse/notifications\nHeaders: Authorization: Bearer \u003Caccess_token\u003E\nReturns: Server-Sent Events stream with formatted user notifications\n\nSends all notifications every 3 seconds\nFixed: Database session leaks, type mismatch, proper cleanup",
                "operationId": "stream_user_notifications_api_v1_sse_notifications_get",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Token"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        },
        "securitySchemes": {
            "HTTPBearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}